+++ clients/python-client/CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## [0.1.0](https://github.com/lumina-ai-inc/chunkr/compare/chunkr-ai-v0.0.50...chunkr-ai-v0.1.0) (2025-06-24)
+
+
+### ⚠ BREAKING CHANGES
+
+* consolidate HTML/markdown generation into single format choice
+
+### Features
+
+* Consolidate HTML/markdown generation into single format choice ([a974f3f](https://github.com/lumina-ai-inc/chunkr/commit/a974f3fbc2bd9158ca052c21a121b479e0eb7613))
+
 ## [0.0.50](https://github.com/lumina-ai-inc/chunkr/compare/chunkr-ai-v0.0.49...chunkr-ai-v0.0.50) (2025-05-22)